requirements.txt
Flask 
pymongo[srv]
python-dotenv
supabase
flask_bcrypt